modifications

+++ app/src/main/java/com/example/edrank_app/models/TopCollege.kt
@@ -5,6 +5,6 @@ data class TopCollege(
     val city: String,
     val name: String,
     val rank: Int,
-    val score: Double,
+    val score: Float,
     val state: String
 )
